pyproject.toml
[tool.poetry]
name = "product-recommendation-api"
version = "0.1.0"
description = "product recommendation api"
authors = ["Victor Thury <victorhauache@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
pandas = "^2.2.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
redis = "^5.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ipykernel = "^6.29.5"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = 'app/app:app'
